+++ src/index.ts
@@ -1,4 +1,4 @@
-/* Copyright(C) 2019-2024, HJD (https://github.com/hjdhjd). All rights reserved.
+/* Copyright(C) 2019-2025, HJD (https://github.com/hjdhjd). All rights reserved.
  *
  * index.ts: UniFi Access API registration.
  */
@@ -7,4 +7,3 @@ export * from "./access-api.js";
 export * from "./access-logging.js";
 export * from "./access-types.js";
 export { FetchError } from "@adobe/fetch";
-