+++ docs/Changelog.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. This project uses [semantic versioning](https://semver.org/).
 
+## 1.3.0 (2025-05-27)
+  * New feature: utility appication, ufa, to allow command-line access to the UniFi Access API.
+  * New feature: initial support for Enterprise Access Hubs.
+  * Improvement: additional typings for recent Access releases.
+  * Housekeeping.
+
 ## 1.2.0 (2024-10-02)
   * Improvement: support for UniFi Access v2.3.
   * Housekeeping.